test: keep loader-free await tests on React
